fix(collapse): rename generateUUID to generate-uuid and update imports #866
