+++ libs/angular-components/tedi/helpers/index.ts
@@ -1,1 +1,1 @@
-export * from "./generateUUID";
+export * from "./generate-uuid";
